Only write a .env if file does not exist and env var exists

+++ scripts/start.sh
@@ -5,7 +5,7 @@ webroot=/www
 cd /www
 chown -Rf www:www *
 
-if [[ -z "${env}" ]]; then
+if [ -z ${env+x}  ] && [ ! -f /www/.env ]; then
     echo $env > .env
 fi
